+++ src/exports/generator/index.js
@@ -1,0 +1,7 @@
+import shape from './shape.js';
+import svg from './svg.js';
+
+export {
+  shape,
+  svg,
+}